Add files via upload
This version of the font was created by David A Roberts and expanded on by jan Same
Version 5 is in the works with all the nimi pi ku lili.
In the meantime here is version 4.9. This version, like version 4.2, uses OpenType features to dynamically compose compound glyphs.
4.9 includes the nimi pi ku suli and a few new alternate glyphs for punctuation marks, n and a.
This version also fixes a few issues with the compound glyphs from version 4.2.